refactor: :recycle: move sqljs code to own folder
