+++ src/sqljs/tests/buildSelectQuery.test.ts
