+++ src/sqljs/tests/buildUpdateQuery.test.ts
