+++ src/sqljs/tests/convertCreateTableStatement.test.ts
