+++ src/sqljs/tests/convertToObjects.test.ts
