+++ src/sqljs/tests/delete.test.ts
