+++ src/sqljs/tests/formatInsertStatement.test.ts
