+++ test.ts
@@ -1,0 +1,3 @@
+import { Sibyl, bSibyl, tSibyl } from './dist/index'
+
+const {} = await Sibyl()
